refactor "emoji to char" function into separate file
